refactor(checker): split indexed access helper methods into submodule
